+++ index.html
@@ -137,7 +137,6 @@
         />
       </div>
     </div>
-    <script src="./script.js"></script>
-    <!-- <script src="./carousel.js"></script> -->
+    <script type="module" src="./app.js"></script>
   </body>
 </html>
